refactor: 패키지 분리 및 금액만큼 로또 생성 테스트 수정
