+++ src/main/kotlin/lotto/Main.kt
@@ -1,5 +1,8 @@
 package lotto
 
+import lotto.view.InputView
+import lotto.view.ResultView
+
 fun main() {
     LottoApplication(InputView, ResultView).run()
 }
