+++ src/main/kotlin/lotto/domain/Rank.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import java.util.function.BiPredicate
