+++ src/test/kotlin/lotto/domain/LottoTest.kt
@@ -1,9 +1,8 @@
-package lotto
+package lotto.domain
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.ValueSource
 import java.lang.IllegalArgumentException
